requirements.txt
# --- Core Framework ---
fastapi==0.115.0
uvicorn==0.31.0

# --- Environment Config ---
python-dotenv==1.0.1

# --- Database (PostgreSQL) ---
SQLAlchemy==2.0.36
psycopg[binary]==3.2.10

# --- Redis (Cache / Queue / Celery backend) ---
redis==5.0.8

# --- Optional (for async ORM, if you plan to use it later) ---
sqlmodel==0.0.22
# asyncpg==0.29.0
aioredis==2.0.1

# --- Optional (for background tasks & scheduling) ---
celery==5.4.0
rq==1.16.2

sentry-sdk
